Log a direct link to the graylog message
A link to the stream is useful, but it's even more useful to be able to get
directly to the message. This is especially true if you are looking at an older
message that's outside of the time window given.

This commit misses one thing: referencing the correct index. Hopefully that can
be fixed before a merge. In the meantime I'm using `graylog_deflector`, which
will the index active at the time of using the URL.